refactor(validation): split 988-line god object into 8 domain modules (#86)
Split src/lib/validation.ts into domain-specific modules under
src/lib/validation/ with barrel re-export preserving all 14 exports.

Fixes #72

Cost-free: ✅
